JOIN 했을 때 채널에 있는 유저들에게 메시지 날림
